created folders and grouped files for the arm, drive, intake, vision, and auto command types
