+++ src/main/java/frc/robot/RobotContainer.java
@@ -8,14 +8,14 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import frc.robot.commands.DefaultDriveUpdatePose;
-import frc.robot.commands.DropCone;
-import frc.robot.commands.DropCube;
-import frc.robot.commands.FloorPickup;
-import frc.robot.commands.FloorDrop;
-import frc.robot.commands.IntakeCone;
-import frc.robot.commands.IntakeCube;
-import frc.robot.commands.SetDriveTarget;
+import frc.robot.commands.drive.DefaultDriveUpdatePose;
+import frc.robot.commands.drive.SetDriveTarget;
+import frc.robot.commands.intake.armIntake.DropCone;
+import frc.robot.commands.intake.armIntake.DropCube;
+import frc.robot.commands.intake.armIntake.IntakeCone;
+import frc.robot.commands.intake.armIntake.IntakeCube;
+import frc.robot.commands.intake.floor.FloorDrop;
+import frc.robot.commands.intake.floor.FloorPickup;
 import frc.robot.subsystems.DrivetrainSubsystem;
 import frc.robot.subsystems.FloorIntake;
 import frc.robot.subsystems.Intake;
