+++ src/main/java/frc/robot/commands/arm/MoveArmToPostion.java
@@ -2,40 +2,40 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.ArmSubsystem;
 
-public class IntakeCone extends CommandBase {
-  /** Creates a new ConeUtil. */
-  private final Intake intake;
-  public IntakeCone(Intake intake) {
+public class MoveArmToPostion extends CommandBase {
+  private ArmSubsystem arm;
+  private double targetPosition;
+  /** Creates a new MoveArmToPostion. */
+  public MoveArmToPostion(ArmSubsystem arm, double targetPosition) {
     // Use addRequirements() here to declare subsystem dependencies.
-    this.intake = intake;
-    addRequirements(this.intake);
+
+    this.arm=arm;
+    this.targetPosition=targetPosition;
+    addRequirements(arm);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+    arm.setTargetPosition(targetPosition);
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-    intake.conePickup(Constants.INTAKE_SPEED);
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    intake.intakeStop();
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return true;
   }
 }
