+++ src/main/java/frc/robot/commands/drive/DefaultDriveUpdatePose.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.drive;
 
 import java.util.function.DoubleSupplier;
