+++ src/main/java/frc/robot/commands/intake/armIntake/DropCone.java
@@ -2,15 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.intake.armIntake;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
 import frc.robot.subsystems.Intake;
 
-public class DropCube extends CommandBase {
-  /** Creates a new DropCube. */
+public class DropCone extends CommandBase {
+  /** Creates a new DropCone. */
   private final Intake intake;
-  public DropCube(Intake intake) {
+  public DropCone(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
     this.intake = intake;
     addRequirements(this.intake);
@@ -23,7 +23,7 @@ public class DropCube extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    intake.cubeDrop(Constants.INTAKE_SPEED);
+    intake.coneDrop(Constants.INTAKE_SPEED);
   }
 
   // Called once the command ends or is interrupted.
