+++ src/main/java/frc/robot/commands/intake/armIntake/IntakeCone.java
@@ -2,46 +2,35 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import java.util.function.DoubleSupplier;
+package frc.robot.commands.intake.armIntake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
-import frc.robot.subsystems.ArmSubsystem;
+import frc.robot.subsystems.Intake;
 
-public class MoveArm extends CommandBase {
-  /** Creates a new moveArm. */
- private ArmSubsystem arm;
- private DoubleSupplier armSpeed;
-
-  public MoveArm(ArmSubsystem arm,DoubleSupplier armSpeed ) {
+public class IntakeCone extends CommandBase {
+  /** Creates a new ConeUtil. */
+  private final Intake intake;
+  public IntakeCone(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
-    this.arm = arm; 
-    this.armSpeed = armSpeed;
-    addRequirements(this.arm);
-
+    this.intake = intake;
+    addRequirements(this.intake);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
+  public void initialize() {}
 
-  }
-    
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    double armPower = armSpeed.getAsDouble()*Constants.ARM_POWER_SCALING;
-    arm.moveArm(armPower);
+    intake.conePickup(Constants.INTAKE_SPEED);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    arm.moveArm(0);
-
+    intake.intakeStop();
   }
 
   // Returns true when the command should end.
