+++ src/main/java/frc/robot/commands/intake/armIntake/IntakeCube.java
@@ -2,18 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.intake.armIntake;
 import frc.robot.Constants;
-import frc.robot.subsystems.FloorIntake;
+import frc.robot.subsystems.Intake;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class FloorDrop extends CommandBase {
-  /** Creates a new FloorIntakeOut. */
-  FloorIntake floorIntake;
-  public FloorDrop(FloorIntake floorIntake) {
+public class IntakeCube extends CommandBase {
+  /** Creates a new CubeUtil. */
+  private final Intake intake;
+  public IntakeCube(Intake intake) {
     // Use addRequirements() here to declare subsystem dependencies.
-    this.floorIntake = floorIntake;
-    addRequirements(this.floorIntake);
+    this.intake = intake;
+    addRequirements(this.intake);
   }
 
   // Called when the command is initially scheduled.
@@ -23,13 +23,13 @@ public class FloorDrop extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    floorIntake.intakeDrop(Constants.FLOOR_INTAKE_SPEED);
+    intake.coneDrop(Constants.INTAKE_SPEED);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    floorIntake.intakeStop();
+    intake.intakeStop();
   }
 
   // Returns true when the command should end.
