+++ src/main/java/frc/robot/commands/intake/floor/FloorDrop.java
@@ -2,15 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-import frc.robot.subsystems.FloorIntake;
+package frc.robot.commands.intake.floor;
 import frc.robot.Constants;
+import frc.robot.subsystems.FloorIntake;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class FloorPickup extends CommandBase {
-  /** Creates a new FloorIntakeIn. */
+public class FloorDrop extends CommandBase {
+  /** Creates a new FloorIntakeOut. */
   FloorIntake floorIntake;
-  public FloorPickup(FloorIntake floorIntake) {
+  public FloorDrop(FloorIntake floorIntake) {
     // Use addRequirements() here to declare subsystem dependencies.
     this.floorIntake = floorIntake;
     addRequirements(this.floorIntake);
@@ -23,7 +23,7 @@ public class FloorPickup extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    floorIntake.intakePickup(Constants.FLOOR_INTAKE_SPEED);
+    floorIntake.intakeDrop(Constants.FLOOR_INTAKE_SPEED);
   }
 
   // Called once the command ends or is interrupted.
