+++ src/main/java/frc/robot/commands/intake/floor/FloorPickup.java
@@ -2,40 +2,39 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-
+package frc.robot.commands.intake.floor;
+import frc.robot.subsystems.FloorIntake;
+import frc.robot.Constants;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.ArmSubsystem;
 
-public class MoveArmToPostion extends CommandBase {
-  private ArmSubsystem arm;
-  private double targetPosition;
-  /** Creates a new MoveArmToPostion. */
-  public MoveArmToPostion(ArmSubsystem arm, double targetPosition) {
+public class FloorPickup extends CommandBase {
+  /** Creates a new FloorIntakeIn. */
+  FloorIntake floorIntake;
+  public FloorPickup(FloorIntake floorIntake) {
     // Use addRequirements() here to declare subsystem dependencies.
-
-    this.arm=arm;
-    this.targetPosition=targetPosition;
-    addRequirements(arm);
+    this.floorIntake = floorIntake;
+    addRequirements(this.floorIntake);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-    arm.setTargetPosition(targetPosition);
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    floorIntake.intakePickup(Constants.FLOOR_INTAKE_SPEED);
+  }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    floorIntake.intakeStop();
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return true;
+    return false;
   }
 }
